Fix File Browser Header Path Navigation Issue (#59)
* Fix File Browser Header Path Navigation Issue

Fixed an issue where navigating to a folder by clicking on a folder name in the Header Path navigation (i.e. `Root/Folder/Names/Here/`), would incorrectly add a `t/` folder to the base of the path. We now correctly chop off all of the leading `Root/` prefix when building paths instead of only `Roo`, fixing this issue.

* typo

---------

Co-authored-by: Alex "mcmonkey" Goodwin <git_commits@alexgoodwin.dev>